# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API Specification
  description: |
    This OpenAPI document describes the WASAPhoto project REST API.

    Keep in touch with your friends by sharing photos of special moments,
    thanks to WASAPhoto! Directly from your PC,
    you can upload your photos, and they will be visible to everyone
    who is following you.

    **Project details at [Project.pdf](http://gamificationlab.uniroma1.it/notes/Project.pdf)**
  version: "1.0.6"
servers:
  - url: http://localhost:8080/
    description: Localhost development server
...
paths:
...
  /users/{userId}/stream:
    description: Resource of photos stream for a user
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: ["follow"]
      operationId: getMyStream
      summary: Get my own stream
      description: |
        Get my own post stream (you are not allowed to see others' stream),
        using cursor pagination.
        It will contain posts from the profiles you follow,
        in reverse chronological order.
      parameters:
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/PaginatedPhotosResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
        "401": { $ref: "#/components/responses/LoginError" }
        "403": { $ref: "#/components/responses/AuthorizationError" }
components:
  parameters:
    UserId:
      name: userId
      required: true
      in: path
      description: The unique ID of a user
      schema: { $ref: "#/components/schemas/ResourceId" }
...
    PageCursor:
      name: pageCursor
      description: |
        In cursor pagination, this refers to
        the ID of the latest resource received by the client.
      required: false
      in: query
      schema: { $ref: "#/components/schemas/ResourceId" }
  responses:
    BadRequest:
      description: |
        The request was not compliant with the documentation
        (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: |
        An unexpected error occurred on the server
        while processing this request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    LoginError:
      description: |
        Authentication is required to perform this action,
        and no *valid* authentication was included in this request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AuthorizationError:
      description: |
        A valid authentication is included in this request,
        but you are not authorized to perform this action.

        For instance, you are trying to modify someone else's data,
        or access to information about someone who banned you.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PaginatedPhotosResult:
      description: |
        The current page was successfully returned
        and it may contain some photos.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PaginatedPhotos" }
...
  schemas:
    Error:
      description: Result representing an error description
      type: string
      minLength: 0
      maxLength: 100
      readOnly: true
    ResourceId:
      description: The unique resource ID as a UUID string
      type: string
      example: "123e4567-e89b-12d3-a456-426614174000"
      minLength: 36
      maxLength: 36
      readOnly: true
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
    Username:
      description: User's personal and unique username
      type: string
      example: john_doe_42
      pattern: "^.*$"
      minLength: 3
      maxLength: 16
...
    User:
      description: |
        Representation of a user, with all the details useful in his/her profile.
        It won't directly include his/her photos, because they must be
        requested separately, using pagination.
      type: object
      properties:
        id: { $ref: "#/components/schemas/ResourceId" }
        name:
          description: User's name
          type: string
          example: John
          pattern: "^.*$"
          minLength: 2
          maxLength: 256
        surname:
          description: User's surname
          type: string
          minLength: 0
          maxLength: 256
          example: Doe
          pattern: "^.*?$"
          default: ""
        username:
          $ref: "#/components/schemas/Username"
        followersCount:
          description: Total count of the followers of a user
          type: integer
          readOnly: true
          minimum: 0
          example: 500
        followingsCount:
          description: Total count of the followings of a user
          type: integer
          readOnly: true
          minimum: 0
          example: 80
        postsCount:
          description: Total count of the posts of a user
          type: integer
          readOnly: true
          minimum: 0
          example: 12
        banned:
          description: |
            It indicates if the user performing this request
            has banned the user represented by this resource.
          type: boolean
          readOnly: true
          example: false
        following:
          description: |
            It indicates whether or not the user performing this request
            is currently following the user represented by this resource.
          type: boolean
          readOnly: true
          example: true
    DateTime:
      description: Standard datetime representation
      type: string
      format: date-time
      example: "2017-07-21T17:32:28Z"
      minLength: 20
      maxLength: 20
      readOnly: true
    StaticImageUrl:
      description: |
        The direct URL to the image file.

        It may also be stored on a different server,
        which can be dedicated to serve static resources,
        or maybe a CDN or may also be simply this server.

        The file can be cached and it's guaranteed to never change.
      type: string
      minLength: 10
      maxLength: 256
      example: https://static.example.com/img_1111_d404401c8c6495b206fc35c95e55a6d5.webp
    Photo:
      description: |
        A photo post, with aggregate information about likes and comments.
        Detailed comments can be requested separately using an appropriate paginated endpoint.
      type: object
      properties:
        id: { $ref: "#/components/schemas/ResourceId" }
        imageUrl: { $ref: "#/components/schemas/StaticImageUrl" }
        author: { $ref: "#/components/schemas/User" }
        publishDate: { $ref: "#/components/schemas/DateTime" }
        likesCount:
          description: Total count of the likes this post received
          type: integer
          minimum: 0
          example: 1200
          readOnly: true
        commentsCount:
          description: Total count of the comments this post received
          type: integer
          minimum: 0
          example: 300
          readOnly: true
        liked:
          description: |
            The current user liked this photo.
            If the request isn't authenticated, this will always be false.
          type: boolean
          readOnly: true
...
    PaginationInfo:
      type: object
      readOnly: true
      description: |
        Additional data necessary to handle
        paginated responses and subsequent requests.
      properties:
        nextPageCursor:
          type: string
          description: String to be sent as pageCursor in the subsequent request
          minLength: 1
          maxLength: 80
        pageData:
          type: array
          minItems: 0
          maxItems: 20
          items:
            type: object
          description: Data of this page, if any
...
    PaginatedPhotos:
      description: Current page of photos
      allOf:
        - $ref: "#/components/schemas/PaginationInfo"
        - description: Current selected page
          type: object
          readOnly: true
          properties:
            pageData:
              description: Photos of the current page
              type: array
              minItems: 0
              maxItems: 20
              items: { $ref: "#/components/schemas/Photo" }
...
  securitySchemes:
    UserIdAuth:
      description: |
        User authentication with the user ID passed
        as it would be an authentication token.
        Of course, that's extremely insecure, but it's done
        according to the project specification.
      type: http
      scheme: bearer

# Apply security scheme globally, disabling it explicitly when unnecessary.
security:
  - UserIdAuth: []